requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
requests
python-dotenv
fastapi
uvicorn
pdfplumber
python-docx